Use ".<modulename>" as default third component for inferred domains.
This is needed because of a problem in the Translations BO page. The system displays a tree of Domain parts, but there's a catch.
Only the last part of the domain is clickable, and any wordings assigned to the branches aren't accessible.

Consider the following two cases:

1) The module contains wordings with only one domain: Modules.Mymodule
2) The module contains wordings with two domains: Modules.Mymodule and Modules.Mymodule.Somesource

In the first case, "Mymodule" is clickable, all wordigns are accessible.
In the second case, only "Somesource is clickable", and wordings stored under "Modules.Mymodule" aren't accessible.

To workaround this problem, the default domain for Mymodule changes from Modules.Mymodule to Modules.Mymodule.Mymodule